some cleanup

+++ src/variables-extern.c
@@ -91,10 +91,10 @@ extern gchar *buffer_dernier_fichier;
 extern gchar *nom_fichier_backup;
 extern gchar *dernier_chemin_de_travail;
 extern gint nb_derniers_fichiers_ouverts;          /* contient le nb de derniers fichiers ouverts */
-extern gint nb_max_derniers_fichiers_ouverts;        /* contient le nb max que peut contenir nb_derniers_fichiers_ouverts ( réglé dans les paramètres ) */
+extern gdouble nb_max_derniers_fichiers_ouverts;        /* contient le nb max que peut contenir nb_derniers_fichiers_ouverts ( réglé dans les paramètres ) */
 extern gchar **tab_noms_derniers_fichiers_ouverts;
-extern gint compression_fichier;
-extern gint compression_backup;
+extern gdouble compression_fichier;
+extern gdouble compression_backup;
 
 
 
@@ -314,7 +314,7 @@ extern GSList *gsliste_echeances;                 /* contient la liste des struc
 extern gint nb_echeances;
 extern gint no_derniere_echeance;
 
-extern gint decalage_echeance;
+extern gdouble decalage_echeance;
 
 extern gint ancienne_largeur_echeances;
